Move commands under fs module, export with single module file
