+++ src/commands/fs/filetypeResolver.js
