+++ src/commands/fs/fs.js
@@ -1,0 +1,15 @@
+import { createCommandLs } from "./commandLs.js";
+import { createCommandUp } from "./commandUp.js";
+import { createCommandCd } from "./commandCd.js";
+import { createCommandCat } from "./commandCat.js";
+import { createCommandAdd } from "./commandAdd.js";
+import { createCommandMkdir } from "./commandMkdir.js";
+
+export {
+    createCommandLs,
+    createCommandUp,
+    createCommandCd,
+    createCommandCat,
+    createCommandAdd,
+    createCommandMkdir
+}
